[12.0][MIG[REF] Split Fiscal Mixin Document in two class to separate Fields and Methods for allow use it in inherit class.
