+++ l10n_br_fiscal/models/__init__.py
@@ -4,6 +4,7 @@ from . import data_abstract
 from . import data_product_abstract
 from . import data_ncm_nbs_abstract
 from . import document_workflow
+from . import document_fiscal_mixin_methods
 from . import document_fiscal_mixin
 from . import document_fiscal_line_mixin_methods
 from . import document_fiscal_line_mixin
